refactor(frontend): extract session history hydration helpers
